# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Smart Dispatch Dashboard API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type VehicleType = typeof VehicleType[keyof typeof VehicleType];


export const VehicleType = {
  fixed: 'fixed',
  outsource: 'outsource',
} as const;

export type VehicleStatus = typeof VehicleStatus[keyof typeof VehicleStatus];


export const VehicleStatus = {
  empty: 'empty',
  busy: 'busy',
  offline: 'offline',
} as const;

export interface Vehicle {
  id: number;
  name: string;
  plate: string;
  type: VehicleType;
  status: VehicleStatus;
  driverName: string;
  phone: string;
  capacity?: number;
  /** @nullable */
  queuePosition: number | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export type VehicleInputType = typeof VehicleInputType[keyof typeof VehicleInputType];


export const VehicleInputType = {
  fixed: 'fixed',
  outsource: 'outsource',
} as const;

export interface VehicleInput {
  name: string;
  plate: string;
  type: VehicleInputType;
  driverName: string;
  phone: string;
  capacity?: number;
  notes?: string;
}

export type VehicleUpdateStatus = typeof VehicleUpdateStatus[keyof typeof VehicleUpdateStatus];


export const VehicleUpdateStatus = {
  empty: 'empty',
  busy: 'busy',
  offline: 'offline',
} as const;

export interface VehicleUpdate {
  name?: string;
  plate?: string;
  status?: VehicleUpdateStatus;
  driverName?: string;
  phone?: string;
  capacity?: number;
  /** @nullable */
  queuePosition?: number | null;
  /** @nullable */
  notes?: string | null;
}

export type TaskType = typeof TaskType[keyof typeof TaskType];


export const TaskType = {
  hotel_pickup: 'hotel_pickup',
  airport_run: 'airport_run',
  extra: 'extra',
} as const;

export type TaskStatus = typeof TaskStatus[keyof typeof TaskStatus];


export const TaskStatus = {
  draft: 'draft',
  assigned: 'assigned',
  in_progress: 'in_progress',
  completed: 'completed',
  cancelled: 'cancelled',
} as const;

export interface Task {
  id: number;
  type: TaskType;
  status: TaskStatus;
  /** @nullable */
  flightCode?: string | null;
  passengerCount: number;
  pickupLocation: string;
  dropoffLocation: string;
  scheduledTime: string;
  /** @nullable */
  actualPickupTime?: string | null;
  /** @nullable */
  actualDropoffTime?: string | null;
  /** @nullable */
  vehicleId?: number | null;
  /** @nullable */
  vehicleName?: string | null;
  /** @nullable */
  driverName?: string | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  fee?: number | null;
  createdAt: string;
}

export type TaskInputType = typeof TaskInputType[keyof typeof TaskInputType];


export const TaskInputType = {
  hotel_pickup: 'hotel_pickup',
  airport_run: 'airport_run',
  extra: 'extra',
} as const;

export interface TaskInput {
  type: TaskInputType;
  flightCode?: string;
  passengerCount: number;
  pickupLocation: string;
  dropoffLocation: string;
  scheduledTime: string;
  vehicleId?: number;
  notes?: string;
  fee?: number;
}

export type TaskUpdateStatus = typeof TaskUpdateStatus[keyof typeof TaskUpdateStatus];


export const TaskUpdateStatus = {
  draft: 'draft',
  assigned: 'assigned',
  in_progress: 'in_progress',
  completed: 'completed',
  cancelled: 'cancelled',
} as const;

export interface TaskUpdate {
  status?: TaskUpdateStatus;
  /** @nullable */
  vehicleId?: number | null;
  flightCode?: string;
  passengerCount?: number;
  pickupLocation?: string;
  dropoffLocation?: string;
  scheduledTime?: string;
  /** @nullable */
  actualPickupTime?: string | null;
  /** @nullable */
  actualDropoffTime?: string | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  fee?: number | null;
}

export interface TaskImport {
  tasks: TaskInput[];
}

export interface ImportResult {
  created: number;
  skipped: number;
  tasks?: Task[];
}

export interface BatchNotify {
  taskIds: number[];
}

export interface BatchNotifyResult {
  sent: number;
  failed: number;
}

export interface TasksSummary {
  total: number;
  draft: number;
  assigned: number;
  inProgress: number;
  completed: number;
  cancelled: number;
  hotelPickups: number;
  airportRuns: number;
  extras: number;
  todayCompleted: number;
  todayRevenue: number;
}

export interface AccountingRecord {
  id: number;
  vehicleId: number;
  /** @nullable */
  vehicleName?: string | null;
  taskId: number;
  amount: number;
  date: string;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export interface VehicleRevenue {
  vehicleId: number;
  vehicleName: string;
  driverName?: string;
  totalRevenue: number;
  tripCount: number;
}

export type WhatsAppWebhookPayloadEntryItem = { [key: string]: unknown };

export interface WhatsAppWebhookPayload {
  entry?: WhatsAppWebhookPayloadEntryItem[];
}

export type ListVehiclesParams = {
type?: ListVehiclesType;
status?: ListVehiclesStatus;
};

export type ListVehiclesType = typeof ListVehiclesType[keyof typeof ListVehiclesType];


export const ListVehiclesType = {
  fixed: 'fixed',
  outsource: 'outsource',
} as const;

export type ListVehiclesStatus = typeof ListVehiclesStatus[keyof typeof ListVehiclesStatus];


export const ListVehiclesStatus = {
  empty: 'empty',
  busy: 'busy',
  offline: 'offline',
} as const;

export type ListTasksParams = {
status?: ListTasksStatus;
type?: ListTasksType;
date?: string;
};

export type ListTasksStatus = typeof ListTasksStatus[keyof typeof ListTasksStatus];


export const ListTasksStatus = {
  draft: 'draft',
  assigned: 'assigned',
  in_progress: 'in_progress',
  completed: 'completed',
  cancelled: 'cancelled',
} as const;

export type ListTasksType = typeof ListTasksType[keyof typeof ListTasksType];


export const ListTasksType = {
  hotel_pickup: 'hotel_pickup',
  airport_run: 'airport_run',
  extra: 'extra',
} as const;

export type ListAccountingRecordsParams = {
vehicleId?: number;
startDate?: string;
endDate?: string;
};