# lib/api-zod/src/generated/types/batchNotifyResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Smart Dispatch Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface BatchNotifyResult {
  sent: number;
  failed: number;
}