# lib/api-zod/src/generated/types/importResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Smart Dispatch Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { Task } from './task';

export interface ImportResult {
  created: number;
  skipped: number;
  tasks?: Task[];
}