# lib/api-zod/src/generated/types/taskImport.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Smart Dispatch Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { TaskInput } from './taskInput';

export interface TaskImport {
  tasks: TaskInput[];
}